+++ Formula/anylint.rb
@@ -1,7 +1,7 @@
 class Anylint < Formula
   desc "Lint anything by combining the power of Swift & regular expressions"
   homepage "https://github.com/Flinesoft/AnyLint"
-  url "https://github.com/Flinesoft/AnyLint.git", :tag => "0.5.0", :revision => "3ef3603e2ae44d8cc8a64a99a7b9af0440aa1b3e"
+  url "https://github.com/Flinesoft/AnyLint.git", :tag => "0.6.0", :revision => "de8f89184529d0bb4995d8a527722e2986d0ec05"
   head "https://github.com/Flinesoft/AnyLint.git"
 
   depends_on :xcode => ["11.4", :build]